+++ tests/testthat/test-list-transpose.R
@@ -4,10 +4,11 @@ test_that("can transpose homogenous list", {
   expect_equal(out, list(a = c(x = 1, y = 3), b = c(x = 2, y = 4)))
 })
 
-test_that("can transpose data frames", {
+test_that("can't transpose data frames", {
   df <- data.frame(x = 1:2, y = 4:5)
-  out <- list_transpose(df)
-  expect_equal(out, list(c(x = 1, y = 4), c(x = 2, y = 5)))
+
+  # i.e. be consistent with other `list_*()` functions from purrr/vctrs
+  expect_snapshot(error = TRUE, list_transpose(df))
 })
 
 test_that("transposing empty list returns empty list", {